\name{RcmdrPlugin.SLC-internal}
\title{RcmdrPlugin.SLC: A GUI for slc-package (internal functions)}
\alias{Rcmdr.help.RcmdrPlugin.SLC}
\alias{Rcmdr.help.SLC}
\alias{Rcmdr.slcestimates}

\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for
  RcmdrPlugin.SLC package. 
  For user information ask for help about \code{\link[RcmdrPlugin.SLC]{RcmdrPlugin.SLC}}.}
\keyword{internal}
