\name{RcmdrPlugin.SLC-package}
\title{RcmdrPlugin.SLC: A GUI for SLC R package}
\alias{RcmdrPlugin.SLC-package}
\alias{RcmdrPlugin.SLC}
\docType{package}
\encoding{UTF-8}
\description{
  This package provides a GUI (Graphical User Interface) for the
  SLC package. It is a Plug-in for Rcmdr (R Commander).

  Functions of \code{SLC} package are available through the
  GUI.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.SLC\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-14-01\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Antonio Solanas <antonio.solanas@ub.edu>, Rumen Manolov <rrumenov13@ub.edu> & Patrick Onghena <patrick.onghena@ped.kuleuven.be>.

Maintainer: Rumen Manolov <rrumenov13@ub.edu>
}

\keyword{package}
\seealso{
For more information see \code{\link[Rcmdr]{Rcmdr-package}} and \code{\link[SLC]{SLC-package}}.
}

